test-tc-shuffle-cls_api.sh: Only fail for specific kernel
This is to avoid catching fixed issue by open issue checker.
The original bug was opened for old kernel with mlnx ofed but the fix
needs to be done in the kernel so the issue is marked as wont fix in
the issue tracker.
The comment was left for history but we dont want to catch it
as current kernels are fixed.
For the specific kernel it fails we now also fail the test.

+++ test-tc-shuffle-cls_api.sh
@@ -3,7 +3,7 @@
 # reload module act_gact before adding rules to get into the request_module flow in gact
 # which will return EAGAIN and expose bug in cls_api module.
 #
-# Bug SW #1294281: [ASAP MLNX OFED] kernel crash adding drop rules when act_gact is not loaded
+# BugSW #1294281: [ASAP MLNX OFED] kernel crash adding drop rules when act_gact is not loaded
 #
 
 NIC=${1:-ens5f0}
@@ -16,6 +16,10 @@ RUNFILE=/tmp/test-$$
 RULE_COUNT=${RULE_COUNT:-100}
 GROUP_COUNT=${GROUP_COUNT:-50}
 ROUND_COUNT=${ROUND_COUNT:-50}
+
+if [ `uname -r` = "3.10.0-327.el7.x86_64" ]; then
+    fail "Bug SW #1294281: [ASAP MLNX OFED] kernel crash adding drop rules when act_gact is not loaded"
+fi
 
 rm -f $TMPFILE
 rm -f $RUNFILE
